fix: move db init and changed tests expectations
